+++ ent/dbase/interface.go
@@ -1,8 +1,0 @@
-package dbase
-
-import "github.com/gnames/gndiff/ent/record"
-
-type DBase interface {
-	Init([]record.Record) error
-	Select(string) ([]record.Record, error)
-}