+++ ent/exact/interface.go
@@ -1,8 +1,0 @@
-package exact
-
-import "github.com/gnames/gndiff/ent/record"
-
-type Exact interface {
-	Init([]record.Record)
-	Find(string) bool
-}