+++ ent/ingester/interface.go
@@ -1,7 +1,0 @@
-package ingester
-
-import "github.com/gnames/gndiff/ent/record"
-
-type Ingester interface {
-	Records(path string) ([]record.Record, error)
-}