+++ main.go
@@ -1,6 +1,5 @@
-The MIT License (MIT)
-
-Copyright © 2021-2023 Dmitry Mozzherin <dmozzherin@gmail.com>
+/*
+Copyright © 2021 Dmitry Mozzherin <dmozzherin@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -19,3 +18,11 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
+*/
+package main
+
+import "github.com/gnames/gndiff/cmd"
+
+func main() {
+	cmd.Execute()
+}