+++ pkg/ent/exact/interface.go
@@ -1,0 +1,8 @@
+package exact
+
+import "github.com/gnames/gndiff/pkg/ent/record"
+
+type Exact interface {
+	Init([]record.Record)
+	Find(string) bool
+}