+++ pkg/ent/ingester/interface.go
@@ -1,0 +1,7 @@
+package ingester
+
+import "github.com/gnames/gndiff/pkg/ent/record"
+
+type Ingester interface {
+	Records(path string) ([]record.Record, error)
+}